+++ projects/core/src/cms/store/actions/cms-group.actions.ts
@@ -1,3 +1,3 @@
-export * from './component.action';
+export * from './components.action';
 export * from './navigation-entry-item.action';
 export * from './page.action';
