+++ projects/core/src/cms/store/selectors/cms-group.selectors.ts
@@ -1,4 +1,4 @@
-export * from './component.selectors';
+export * from './components.selectors';
 export * from './feature.selectors';
 export * from './navigation-entry-item.selectors';
 export * from './page.selectors';
