+++ projects/core/src/cms/store/selectors/components.selectors.ts
